+++ src/js/language/cs_CZ.js
@@ -238,6 +238,7 @@
                 country: 'Prosím zadejte správné telefoní číslo pro %s',
                 countries: {
                     AE: 'Spojené arabské emiráty',
+                    BG: 'Bulharsko',
                     BR: 'Brazílii',
                     CN: 'Čínu',
                     CZ: 'Českou Republiku',